+++ plugins/gui/include/gui/selection_details_widget/selection_details_widget.h
@@ -338,6 +338,8 @@ namespace hal
          */
         void selectionToModuleAction(int actionCode);
 
+        void showNoSelection();
+
         QSplitter*           mSplitter;
         SelectionTreeView*   mSelectionTreeView;
         unsigned int         mNumberSelectedItems;